\name{pajengr-package}
\alias{pajengr-package}
\alias{pajengr}
\docType{package}
\title{
  An R Package to encapsulate the PajeNG framework
}
\description{
  The pajengr package contains a set of
  data-processing functions that encapsulate the
  pajeng simulator and the pj_dump functionality
  to get data frames directly from paje trace files.
}
\details{
  Forthcoming.
}
\author{
  Lucas Mello Schnorr

  Maintainer: Lucas M. Schnorr <schnorr@inf.ufrgs.br>
}
\references{
  Forthcoming.
}
\keyword{ package }
\seealso{
  Forthcoming.
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
